Moved load children callback to own file
